pyproject.toml
[tool.poetry]
name = "shering-api"
version = "0.1.0"
description = ""
authors = ["BazaroZero <finatov05@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
djangorestframework-simplejwt = "^5.2.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
